+++ changes.md
@@ -5,6 +5,7 @@
   * Add tests for SYB once
   * Split it to a library and adapter modules
   * Simplify adapter interface
+  * Capture stdout and stderr of test runs.
 
 ## [0.1.3.0] (Rahul Gopinath)
   * Migrate mucheck to Test namespace